backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
Pillow
python-jose